Move lagrange_interpolation
Since this module is only use in advection library, i move it in
advection directory. Lagrange interpolation depends on
sll_boundary_condition_descriptors so now advection depends on it too.
We need to check why there are two versions of lagrange interpolation
in selalib. In advection dir and in lagrange_interpolation dir.